Cpplint

+++ modules/bm_sim/src/packet.cpp
@@ -18,12 +18,14 @@
  *
  */
 
+#include <algorithm>  // for swap
+
 #include "bm_sim/packet.h"
 
 #include "xxhash.h"
 
 void
-Packet::update_signature(unsigned long long seed) {
+Packet::update_signature(uint64_t seed) {
   signature = XXH64(buffer.start(), buffer.get_data_size(), seed);
 }
 
@@ -31,17 +33,16 @@ void
 Packet::set_ingress_ts() {
   ingress_ts = clock::now();
   ingress_ts_ms = std::chrono::duration_cast<std::chrono::milliseconds>(
-    ingress_ts.time_since_epoch()
-  ).count();
+    ingress_ts.time_since_epoch()).count();
 }
 
 Packet::Packet() {
   assert(phv_pool);
-  phv = phv_pool->get(); // needed ?
+  phv = phv_pool->get();  // needed ?
 }
 
 Packet::Packet(int ingress_port, packet_id_t id, packet_id_t copy_id,
-	       int ingress_length, PacketBuffer &&buffer)
+               int ingress_length, PacketBuffer &&buffer)
   : ingress_port(ingress_port), packet_id(id), copy_id(copy_id),
     ingress_length(ingress_length), buffer(std::move(buffer)) {
   assert(phv_pool);
@@ -51,7 +52,7 @@ Packet::Packet(int ingress_port, packet_id_t id, packet_id_t copy_id,
 }
 
 Packet::Packet(int ingress_port, packet_id_t id, packet_id_t copy_id,
-	       int ingress_length, PacketBuffer &&buffer, const PHV &src_phv)
+               int ingress_length, PacketBuffer &&buffer, const PHV &src_phv)
   : ingress_port(ingress_port), packet_id(id), copy_id(copy_id),
     ingress_length(ingress_length), buffer(std::move(buffer)) {
   assert(phv_pool);
@@ -60,11 +61,11 @@ Packet::Packet(int ingress_port, packet_id_t id, packet_id_t copy_id,
   phv = phv_pool->get();
   phv->copy_headers(src_phv);
 }
-  
+
 Packet::~Packet() {
   assert(phv);
   // corner case: phv_pool has already been destroyed
-  if(phv_pool) {
+  if (phv_pool) {
     phv->reset();
     phv->reset_header_stacks();
     phv_pool->release(std::move(phv));
@@ -74,7 +75,7 @@ Packet::~Packet() {
 Packet
 Packet::clone(packet_id_t new_copy_id) const {
   Packet pkt(ingress_port, packet_id, new_copy_id, ingress_length,
-	     buffer.clone(buffer.get_data_size()));
+             buffer.clone(buffer.get_data_size()));
   pkt.phv->copy_headers(*phv);
   return pkt;
 }
@@ -82,8 +83,8 @@ Packet::clone(packet_id_t new_copy_id) const {
 Packet
 Packet::clone_and_reset_metadata(packet_id_t new_copy_id) const {
   Packet pkt(ingress_port, packet_id, new_copy_id, ingress_length,
-	     buffer.clone(buffer.get_data_size()));
-  // TODO: optimize this
+             buffer.clone(buffer.get_data_size()));
+  // TODO(antonin): optimize this
   pkt.phv->copy_headers(*phv);
   pkt.phv->reset_metadata();
   return pkt;
@@ -92,7 +93,7 @@ Packet::clone_and_reset_metadata(packet_id_t new_copy_id) const {
 Packet
 Packet::clone_no_phv(packet_id_t new_copy_id) const {
   Packet pkt(ingress_port, packet_id, new_copy_id, ingress_length,
-	     buffer.clone(buffer.get_data_size()));
+             buffer.clone(buffer.get_data_size()));
   return pkt;
 }
 
@@ -154,7 +155,7 @@ Packet::PHVPool::PHVPool(const PHVFactory &phv_factory)
 std::unique_ptr<PHV>
 Packet::PHVPool::get() {
   std::unique_lock<std::mutex> lock(mutex);
-  if(phvs.size() == 0) {
+  if (phvs.size() == 0) {
     lock.unlock();
     return phv_factory.create();
   }
